# perf: proposition optim core web vitals