--- a/.talismanrc
+++ b/.talismanrc
@@ -1,4 +1,6 @@
 fileignoreconfig:
+- filename: ui/components/Page/components/Footer.tsx
+  checksum: 3892ab85ca63a683c00c8316fd13502a4f0d4835aef111dba0bdcc03fee2c101
 - filename: ui/components/Televersement/DocumentsActionButtons.tsx
   checksum: 766028bfe6d530cbd170ff5af22d644b0e524e8dbac78cfed637a8115694bb9d
 - filename: yarn.lock