--- a/ui/components/Page/SimplePage.tsx
+++ b/ui/components/Page/SimplePage.tsx
@@ -35,6 +35,7 @@ function SimplePage({ title, children }: Props) {
         right="40px"
         bg="bluefrance"
         borderRadius="200px"
+        aria-label="Accéder à la FAQ"
       >
         <Interrogation color="white" pr="0.1em" pb="0.2em" />
       </Link>