--- a/ui/styles/globals.css
+++ b/ui/styles/globals.css
@@ -1,3 +1,8 @@
+@import "remixicon/fonts/remixicon.css";
+@import "react-phone-input-2/lib/style.css";
+@import "react-datepicker/dist/react-datepicker.css";
+@import "./MonthSelect.css";
+
 th {
   text-align: left;
 }