> Cette PR améliore les Core Web Vitals, l’accessibilité et le SEO de la page /operateurs-publics, avec +11 pts de perf et +18 pts SEO.

Bonjour l’équipe TBA,

Je vous ai récemment envoyé ma candidature via Welcome to the Jungle pour le poste d’**Expert Développeur·se Full-Stack (Node, React, MongoDB)** : [Offre Welcome to the Jungle](https://www.welcometothejungle.com/fr/companies/communaute-beta-gouv/jobs/expert-developpeur-se-fullstack-node-react-mongodb-mern-experimente-e_paris)  

---

Dans ma lettre, je vous faisais part de ma curiosité et de mes idées d’amélioration après avoir exploré votre code source. Plutôt que d’en rester au discours, j’ai voulu apporter ma petite pierre à l’édifice :

1. **Fork de votre repo**  
2. **Identification de plusieurs axes d’optimisation**  
   - Performance  
   - SEO  
   - Accessibilité  
3. **Implémentation de mes premières propositions**  
   - Page : `/operateurs-publics` (exemple concret)

---

> **Résumé chiffré des gains Lighthouse (avant vs après)**  
> _Tous les détails et explications se trouvent dans les commentaires de cette PR._

<table>
  <tr>
    <td align="center">
      Avant optimisations<br />
      <img width="807" alt="Avant optimisations Lighthouse"
           src="https://github.com/user-attachments/assets/e32b46c1-4468-4dab-99a2-3ff271e7b1a7" />
    </td>
    <td align="center">
      Après optimisations<br />
      <img width="818" alt="Après optimisations Lighthouse"
           src="https://github.com/user-attachments/assets/17775cd9-bd8d-4175-acd3-72f865769a8d" />
    </td>
  </tr>
  <tr>
    <td colspan="2" align="center">
      <em>Captures d’écran Lighthouse (Performance / Accessibilité / Bonnes pratiques / SEO)</em>
    </td>
  </tr>
</table>

<table>
  <tr>
    <td align="center">
      Avant détails<br />
      <img width="648" alt="Avant détails métriques"
           src="https://github.com/user-attachments/assets/31d36b35-01b4-4e2f-b3a3-d337976bce13" />
    </td>
    <td align="center">
      Après détails<br />
      <img width="644" alt="Après détails métriques"
           src="https://github.com/user-attachments/assets/87f68e5c-0520-4f21-b476-0a2c7a1f64c1" />
    </td>
  </tr>
  <tr>
    <td colspan="2" align="center">
      <em>Détails des métriques clés</em>
    </td>
  </tr>
</table>

---

### 📊 Gains globaux Lighthouse sur `/operateurs-publics`

| 🚩 Audit            | Avant | Après | Évolution   |
|---------------------|:-----:|:-----:|:-----------:|
| **Performance**     |  87   |  98   | **+11 pts** |
| **Accessibilité**   |  92   | 100   | **+ 8 pts** |
| **Bonnes pratiques**|  93   |  96   | **+ 3 pts** |
| **SEO**             |  82   | 100   | **+ 18 pts**|

> **Détails clés :**  
> - **LCP** : 1,5 s → 0,9 s (**– 0,6 s**)  
> - **TBT** : 130 ms → 110 ms (**– 20 ms**)  
> - **CLS** : 0,15 → 0 (**– 0,15**)  

*Ces visuels et chiffres démontrent l’impact concret des optimisations : une interface plus rapide, plus stable et plus accessible.*

---

J’espère que cette petite contribution :  
- **Éveillera** votre intérêt pour mon profil  
- **Rassurera** quant à mes compétences techniques  
- **Illustrera** ma méthodologie de travail  

Je serais également ravi de discuter des autres pistes d’amélioration que j’ai repérées.

J’espère que vous y trouverez quelques idées utiles pour la suite.  
N’hésitez pas à me faire part de vos retours, je serai enchanté de les recevoir !

Bien à vous,

**Florian Sahbi**  
06 10 65 89 29  
florian.sahbi@gmail.com
[LinkedIn](https://www.linkedin.com/in/floriansahbi/) · [GitHub](https://github.com/FlorianSahbi)

### 📊 Annexes

#### Avant optimisations (CLS perceptible)  
Contenu déplacé durant le chargement (images et police lentes → CLS perceptible)  

https://github.com/user-attachments/assets/482f9d47-ddee-40f4-bbbc-a5e21618bb8a

---

#### Après optimisations (CLS éliminé)  
Police et images chargent plus vite → contenu stable, plus de décalage de mise en page  

https://github.com/user-attachments/assets/228e0787-4e6e-4839-bedb-cb3fae9aaded

---

#### Prototype initial (aperçu UI)  
Avant le passage éclair en mode responsive.

https://github.com/user-attachments/assets/a7139291-3328-4efc-8c57-d805e2823bb5

---

#### Prototype actuel (UI amélioré)  
Apres, rendu plus fluide et visuel, quelques optimisations restent encore nécessaires, mais Lighthouse ne sanctionne plus les tests mobiles.

https://github.com/user-attachments/assets/27f8f6bb-25f3-4988-a468-bdb02bf2d9fd